fix FA for arm

+++ backends/npu/custom_op/fused_attention_npu.cc
@@ -148,7 +148,7 @@ std::vector<paddle::Tensor> npu_flash_attention(
            "The value tensor dtype must be bfloat16 or float16 , but got ",
            value_dtype);
 
-  const int32_t head_num = query_tensor_dims[2];
+  const int64_t head_num = query_tensor_dims[2];
   const double scale = 1.0f / std::sqrt(query_tensor_dims[3]);
 
   void* realShiftOptional = nullptr;
